refactor: dissolve ozma_terminal crate into src/ (surface, spawn, cells_for)
