+++ Cargo.toml
@@ -21,11 +21,11 @@ debug = ["ozma_webview/debug"]
 
 [dependencies]
 ab_glyph = "0.2"
+anyhow = { workspace = true }
 arboard = { version = "3", features = ["wayland-data-control"] }
 bevy = { workspace = true }
 bevy_cef = "0.11"
 bevy_cef_core = { workspace = true }
-ozma_terminal = { path = "crates/ozma_terminal" }
 ozma_tty_renderer = { path = "crates/ozma_tty_renderer" }
 ozma_tty_engine = { path = "crates/ozma_tty_engine" }
 ozma_webview = { path = "crates/ozma_webview" }
