--- crates/ozma_terminal/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "ozma_terminal"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-readme.workspace = true
-authors.workspace = true
-publish.workspace = true
-
-[dependencies]
-anyhow            = { workspace = true }
-arboard           = { version = "3", features = ["wayland-data-control"] }
-bevy              = { workspace = true }
-open              = { workspace = true }
-ozma_tty_engine   = { path = "../ozma_tty_engine" }
-ozma_tty_renderer = { path = "../ozma_tty_renderer" }
-tracing           = { workspace = true }
-
-[lints]
-workspace = true
