+++ src/input/mouse/wheel.rs
@@ -269,9 +269,9 @@ mod tests {
     use crate::input::mouse::test_support::{
         CapturedEffects, add_effect_capture_observers, set_phys_cursor, test_metrics,
     };
+    use crate::surface::OzmaTerminal;
     use bevy::input::mouse::MouseScrollUnit;
     use bevy::ui::{ComputedNode, UiGlobalTransform};
-    use ozma_terminal::OzmaTerminal;
     use ozma_tty_engine::TerminalHandle;
     use ozma_tty_renderer::schema::TerminalGrid;
