+++ src/main.rs
@@ -8,6 +8,7 @@ mod configs;
 mod font;
 mod input;
 mod mode;
+mod surface;
 mod surface_geom;
 mod system_set;
 mod theme;
@@ -21,6 +22,7 @@ use crate::clipboard::ClipboardPlugin;
 use crate::input::focus::FocusSyncPlugin;
 use crate::input::hyperlink::HyperlinkInputPlugin;
 use crate::mode::AppMode;
+use crate::surface::SurfacePlugin;
 use crate::window_title::WindowTitlePlugin;
 use bevy::prelude::*;
 use bootstrap::OzmuxBootstrapPlugin;
@@ -32,7 +34,6 @@ use input::ime::ImePlugin;
 use mode::default::DefaultModePlugin;
 use mode::default::DefaultWebviewPointerPlugin;
 use mode::tmux::OzmuxTmuxPlugin;
-use ozma_terminal::OzmaTerminalPlugin;
 use ozma_tty_engine::TerminalHandlePlugin;
 use ozma_tty_renderer::TerminalRendererPlugin;
 use ozma_webview::{OzmaWebviewPlugin, cef_plugin};
@@ -65,10 +66,10 @@ fn main() {
         ))
         .insert_state(initial_mode)
         .add_plugins((
-            OzmaTerminalPlugin {
+            SurfacePlugin,
+            DefaultModePlugin {
                 config_shell: pre_configs.ozma.shell.clone(),
             },
-            DefaultModePlugin,
             ClipboardPlugin,
             TerminalHandlePlugin,
             TerminalRendererPlugin,
